Create logistic_regression_from_scratch.py
First it calculates the log odds which signifies how many more times an event is likely to occur as compared to not occurring. The log odds are simply the dot product of the features (i.e. hours studied) and coefficients added with the intercept. This is passed into the sigmoid function (1/ 1+ e^-z) which calculates the probability. Now, if the probability is greater than the threshold then add 1 to the np array otherwise add 0. Return the np.array